story: refactor two functions to be reusable in lib folder.
