+++ client/src/components/RecipeDetailHeader.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import detectTime from '../lib/detectTime';
 
 const RecipeDetailHeader = (props) => {
   const {
@@ -22,16 +23,6 @@ const RecipeDetailHeader = (props) => {
     categoriesId,
   } = props;
 
-  const detectTime = (time) => {
-    switch (time) {
-      case time === 120:
-        return '< 2 hrs';
-      case time === 999:
-        return '2 hrs >';
-      default:
-        return `< ${time} min`;
-    }
-  };
   return (
     <div className='RecipeDetail-container'>
       <div className='RecipeDetail-photoContainer'>
